Added Account suspension Protection

+++ spring/src/main/java/com/example/controllers/ManagerController.java
@@ -34,6 +34,9 @@ public class ManagerController {
         if (!managerById.get().getRole().equals(UserModel.MANAGER))
             return new ResponseModelListPayload<ExpenseModel>(ResponseModel.ROLE_CHANGED, "Your role has changed. Please login again", null);
 
+        if(!managerById.get().isActive())
+            return new ResponseModelListPayload<>(ResponseModel.INACTIVE, "You account is currently suspended.", null);
+
         return expenseService.getAllExpenses();
     }
 
@@ -47,6 +50,9 @@ public class ManagerController {
         // Check role validity
         if (!managerById.get().getRole().equals(UserModel.MANAGER))
             return new ResponseModelSinglePayload<ExpenseModel>(ResponseModel.ROLE_CHANGED, "Your role has changed. Please login again", null);
+
+        if(!managerById.get().isActive())
+            return new ResponseModelSinglePayload<>(ResponseModel.INACTIVE, "You account is currently suspended.", null);
 
         // If everything is valid, get expense
         return expenseService.getExpense(expenseId);
@@ -67,6 +73,8 @@ public class ManagerController {
         if (!managerById.get().getRole().equals(UserModel.MANAGER))
             return new ResponseModelSinglePayload<ExpenseModel>(ResponseModel.ROLE_CHANGED, "Your role has changed. Please login again", null);
 
+        if(!managerById.get().isActive())
+            return new ResponseModelSinglePayload<>(ResponseModel.INACTIVE, "You account is currently suspended.", null);
         // If everything checks out, update the expense
         return expenseService.updateExpense(expenseId, expenseModelToUpdate, false);
     }
@@ -81,6 +89,8 @@ public class ManagerController {
         if (!managerById.get().getRole().equals(UserModel.MANAGER))
             return new ResponseModel(ResponseModel.ROLE_CHANGED, "Your role has changed. Please login again.");
 
+        if(!managerById.get().isActive())
+            return new ResponseModel(ResponseModel.INACTIVE, "You account is currently suspended.");
         // If everything is valid, delete expense
         return expenseService.deleteExpenseById(expenseId);
     }
@@ -97,6 +107,8 @@ public class ManagerController {
         if(!managerById.get().getRole().equals(UserModel.MANAGER))
             return new ResponseModel(ResponseModel.ROLE_CHANGED, "Your role has changed. Please login again");
 
+        if(!managerById.get().isActive())
+            return new ResponseModel(ResponseModel.INACTIVE, "You account is currently suspended.");
         try {
             expenseService.storeReceiptImage(receiptImage, expenseId);
         } catch (IOException e) {
